requirements.txt
# Palantiny Chatbot Server
# FastAPI 및 ASGI 서버
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# 데이터베이스 (PostgreSQL 비동기)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Redis (4.x+ async 지원)
redis>=5.0.0

# 설정 및 환경변수
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM API (OpenAI, Mock 모드 대비)
openai>=1.12.0
